Change typechecked to consistently return exact input type
Changes typechecked to consistently return the original input string
when passed a string module name as its argument. This replaces the
behavior of sometimes returning the resolved module for that name and
sometimes returning the original string depending on various factors,
which made the return type awkward to actually rely on anyway.

Fixes #97.